Add command-line option to ignore comments
With '-i' or '--ignore-comments', pinpoint will not show comments.

This is useful when you have to present a PDF version on a machine
without pinpoint (of which there are a shocking number).  The default
PDF output includes the comments as notes on interstitial pages, which
breaks up the flow and is possibly embarrassing.

It is also possible to use -i in speaker mode, in the strange case
that you want to see imminent slides and the time but not your notes.

Signed-off-by: Douglas Bagnall <douglas@paradise.net.nz>